move SequenceSegment (and related types) out of Block.cs
